perf: lazy-load rare sync recovery paths
